refactor: extract the RoomWakeLauncher lane from DispatchOperations
Second launching lane peeled. The room-wake (chat) lane — a human message to
a spec's room while no run is live wakes the agent as a read-only room-role
run that answers in the room — is a thin lane over the three seams: reserve
through RunReservation (empty repo set), launch through RunLauncher, and
capture the RoomCommitGuard baseline so a worktree-writing chat surfaces.

Move startRoomRun and its room-only helpers (engagedSessionId,
resumableSessionId) into RoomWakeLauncher over seven collaborators. The
message store is late-bound via useMessages, so it is injected as a
Supplier<MessageStore> read at wake time. seedRoomDelivery is a trivial
run-store wrapper shared with the build and invite lanes, so RoomWakeLauncher
carries its own copy (as the prior extractions did with runBookkeeping);
DispatchOperations keeps its copy for those lanes. DispatchOperations keeps
startRoomRun as a one-line delegator, so the wake seam (ServerStartCommand,
RoomWakeReactor) and SailOperations are unchanged.

Behavior-preserving — RoomWakeLaunchTest passes unchanged, plus three added
cases (no agent block, no run store, launch failure releases the reservation)
that close the lane's defensive guards, so the new api class meets the 100%
coverage gate with no exclude. mvn clean verify green.
